+++ UnitTests/BraintreePayPalTests/BTPayPalVaultRequest_Tests.swift
@@ -30,6 +30,25 @@ class BTPayPalVaultRequest_Tests: XCTestCase {
         let request = BTPayPalVaultRequest()
         XCTAssertEqual(request.paymentType, .vault)
     }
+    
+    // MARK: - userActionAsString
+
+    func testUserActionAsString_whenUserActionNotSpecified_returnsEmptyString() {
+        let request = BTPayPalVaultRequest()
+        XCTAssertEqual(request.userAction.stringValue, "")
+    }
+
+    func testUserActionAsString_whenUserActionIsDefault_returnsEmptyString() {
+        let request = BTPayPalVaultRequest()
+        request.userAction = .none
+        XCTAssertEqual(request.userAction.stringValue, "")
+    }
+
+    func testUserActionAsString_whenUserActionIsSetupNow_returnsSetupNow() {
+        let request = BTPayPalVaultRequest()
+        request.userAction = .setupNow
+        XCTAssertEqual(request.userAction.stringValue, "setup_now")
+    }
 
     // MARK: - parametersWithConfiguration
 
@@ -276,8 +295,75 @@ class BTPayPalVaultRequest_Tests: XCTestCase {
             enablePayPalAppSwitch: true,
             userAuthenticationEmail: "fake@gmail.com"
         )
-
+        
         XCTAssertEqual(request.userAuthenticationEmail, "fake@gmail.com")
         XCTAssertTrue(request.enablePayPalAppSwitch)
+    }
+    
+    func testParameters_enableAppSwitch_withRecurringBillingDetails_returnsAllParams() {
+        let billingPricing = BTPayPalBillingPricing(
+            pricingModel: .autoReload,
+            amount: "test-price",
+            reloadThresholdAmount: "test-threshold"
+        )
+        
+        let billingCycle = BTPayPalBillingCycle(
+            isTrial: false,
+            numberOfExecutions: 12,
+            interval: .month,
+            intervalCount: 13,
+            sequence: 9,
+            startDate: "test-date",
+            pricing: billingPricing
+        )
+
+        let recurringBillingDetails = BTPayPalRecurringBillingDetails(
+            billingCycles: [billingCycle],
+            currencyISOCode: "test-currency",
+            totalAmount: "test-total",
+            productName: "test-product-name",
+            productDescription: "test-product-description",
+            productQuantity: 1,
+            oneTimeFeeAmount: "test-fee",
+            shippingAmount: "test-shipping",
+            productAmount: "test-price",
+            taxAmount: "test-tax"
+        )
+        
+        let request = BTPayPalVaultRequest(enablePayPalAppSwitch: true)
+        request.recurringBillingDetails = recurringBillingDetails
+        request.recurringBillingPlanType = .subscription
+                
+        guard let parameters = try? request.encodedPostBodyWith(configuration: configuration, isPayPalAppInstalled: true, universalLink: URL(string: "some-url")).toDictionary() else {
+            XCTFail()
+            return
+        }
+        
+        guard let planTypeParameter = parameters["plan_type"] as? String else { XCTFail(); return }
+        XCTAssertEqual(planTypeParameter, "SUBSCRIPTION")
+        
+        guard let planMetadata = parameters["plan_metadata"] as? [String: Any] else { XCTFail(); return }
+        XCTAssertEqual(planMetadata["currency_iso_code"] as! String, "test-currency")
+        XCTAssertEqual(planMetadata["name"] as! String, "test-product-name")
+        XCTAssertEqual(planMetadata["product_description"] as! String, "test-product-description")
+        XCTAssertEqual(planMetadata["product_quantity"] as! Int, 1)
+        XCTAssertEqual(planMetadata["one_time_fee_amount"] as! String, "test-fee")
+        XCTAssertEqual(planMetadata["shipping_amount"] as! String, "test-shipping")
+        XCTAssertEqual(planMetadata["product_price"] as! String, "test-price")
+        XCTAssertEqual(planMetadata["tax_amount"] as! String, "test-tax")
+        XCTAssertEqual(planMetadata["total_amount"] as! String, "test-total")
+
+        guard let billingCycles = planMetadata["billing_cycles"] as? [[String:Any]] else { XCTFail(); return }
+        XCTAssertEqual(billingCycles[0]["billing_frequency"] as! Int, 13)
+        XCTAssertEqual(billingCycles[0]["billing_frequency_unit"] as! String, "MONTH")
+        XCTAssertEqual(billingCycles[0]["number_of_executions"] as! Int, 12)
+        XCTAssertEqual(billingCycles[0]["sequence"] as! Int, 9)
+        XCTAssertEqual(billingCycles[0]["start_date"] as! String, "test-date")
+        XCTAssertFalse(billingCycles[0]["trial"] as! Bool)
+        
+        guard let pricingScheme = billingCycles[0]["pricing_scheme"] as? [String:String] else { XCTFail(); return }
+        XCTAssertEqual(pricingScheme["pricing_model"], "AUTO_RELOAD")
+        XCTAssertEqual(pricingScheme["price"], "test-price")
+        XCTAssertEqual(pricingScheme["reload_threshold_amount"], "test-threshold")
     }
 }